+++ Classes/CustomControls/UIImageView+loadProgress.m
@@ -44,7 +44,7 @@
     
     [operation setCompletionBlockWithSuccess: successDownload failure: failDownload];
     
-    [operation setDownloadProgressBlock:^(NSUInteger bytesRead, long long totalBytesRead, long long totalBytesExpectedToRead) {
+    [operation setDownloadProgressBlock:^(NSUInteger bytesRead, NSInteger totalBytesRead, NSInteger totalBytesExpectedToRead) {
         progess.progress = (float)totalBytesRead/(float)totalBytesExpectedToRead;
     }];
